requirements.txt
analiticcl
dataclasses-json
elucidate-client
icecream
lxml
nltk
numpy
scipy
pagexml
python-dateutil==2.8.1
requests
textrepo-client
xmltodict